Display a modal progress dialog while wave export is in progress
Because exporting works through the same stream pipeline that handles all playback and playback settings, only one or the other can be active. To prevent such a situation from happening, the GUI must be blocked while export is in progress, which is done through the means of the added modal dialog. The modal dialog can be universally used for all actions that need to run exclusive of user interaction.